weigthUI 변경, 각종 버그 수정

+++ GSO_Client/Assets/Scripts/Inventory/Data/ItemData.cs
@@ -79,7 +79,7 @@ public class ItemData
     [Header("임시 사용변수")]
     //임시변수(아이템 코드를 통해 데이터베이스에서 불러오기가 가능할때까지)
     public string item_name;
-    public float item_weight; //아이템의 무게
+    public double item_weight; //아이템의 무게
     public ItemType item_type;
     public int item_string_value;
     public int item_purchase_price;
